server/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-dotenv
openai
pydantic
httpx
httpcore
starlette
websockets
watchfiles
typing_extensions
jiter